Splits byzantine_ledger.rs into smaller files
